requirements.txt
torch
transformers
pillow
opencv-python
matplotlib
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git